etl/requirements.txt
pgvector
psycopg2-binary 
sentence-transformers
torch-directml
pandas
mlflow